we are definately ok with these values, so cast off the converison err.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@61930 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/win32/sockets.c
@@ -310,8 +310,8 @@ APR_DECLARE(apr_status_t) apr_connect(apr_socket_t *sock, apr_sockaddr_t *sa)
             tvptr = NULL;
         }
         else {
-            tv.tv_sec = sock->timeout / APR_USEC_PER_SEC;
-            tv.tv_usec = sock->timeout % APR_USEC_PER_SEC;
+            tv.tv_sec =  (long)(sock->timeout / APR_USEC_PER_SEC);
+            tv.tv_usec = (long)(sock->timeout % APR_USEC_PER_SEC);
             tvptr = &tv;
         }
         rc = select(FD_SETSIZE+1, NULL, &wfdset, &efdset, tvptr);
